requirements.txt
pytest==8.1.1
requests==2.32.4
playwright==1.48.0
pytest-playwright==0.7.1
python-dotenv==1.0.1

black==24.3.0
isort==5.13.2